requirements.txt
spire.pdf
pdfminer.six